requirements.txt
Flask==3.0.3
python-escpos==3.1
qrcode==7.4.2
gunicorn==21.2.0
websocket-client
requests
pyusb==1.2.1
evdev
